+++ company/views.py
@@ -15,19 +15,6 @@ from company import forms, helpers
 import core.mixins
 
 
-class SubmitFormOnGetMixin:
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        data = self.request.GET or {}
-        if data:
-            kwargs['data'] = data
-        return kwargs
-
-    def get(self, request, *args, **kwargs):
-        return super().post(request, *args, **kwargs)
-
-
 class CompanyProfileMixin:
 
     @cached_property
@@ -35,7 +22,9 @@ class CompanyProfileMixin:
         return helpers.get_company_profile(self.kwargs['company_number'])
 
 
-class CompanySearchView(SubmitFormOnGetMixin, CountryDisplayMixin, FormView):
+class CompanySearchView(
+    core.mixins.SubmitFormOnGetMixin, CountryDisplayMixin, FormView
+):
     template_name = 'company-search-results-list.html'
     form_class = forms.CompanySearchForm
     page_size = 10